+++ backend/src/main/java/com/mapbefine/mapbefine/location/dto/CoordinateRequest.java
@@ -1,9 +1,7 @@
 package com.mapbefine.mapbefine.location.dto;
 
-import java.math.BigDecimal;
-
 public record CoordinateRequest(
-        BigDecimal latitude,
-        BigDecimal longitude
+        double latitude,
+        double longitude
 ) {
 }
